Fixed an issue with the Delete User test
